refactor PointsLayerPanel into discrete component
